codelens_ai/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
chromadb==1.0.20
langchain==0.3.27
langchain-openai==0.3.32
langchain-community==0.3.27
pytest==8.4.1
pytest-asyncio==1.1.0
